Added functionality for Director AI to command the ghosts to specified locations.
Added some SFX.

Added low tension activities (a random ghost will head to a location close to the player if the tension is on 0 for too long. This is signaled to the player through a sound cue)

Ghosts will now clear the TargetActor value in their behavior blackboard when scattering.