TargetOrbMemory now properly stores the knowledge state about all orbs that were seen so far. State goes back to unexplored after a certain amount of time.
Added EQS node for checking on orbs. Doesn't work yet as the code, triggering the branch in the BT isn't complete yet.

In Hard Mode the individual enemy will now gain speed for every known orb that is collected.

Fixed a bug where the enemy would freeze, when a MazeZone is cleared.

ToDo: Move the TargetOrbMemory component into the controller BP instead of the Pawn's BP